@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<RootNamespace>CloudNimble.SimpleMessageBus.Core</RootNamespace>
 		<AssemblyName>CloudNimble.SimpleMessageBus.Core</AssemblyName>
-		<TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0</TargetFrameworks>
+		<TargetFrameworks>netstandard2.0;net9.0;net8.0;net6.0</TargetFrameworks>
 		<DocumentationFile>$(DocumentationFile)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
 
@@ -15,29 +15,29 @@
 			the core objects necessary to construct a Domain event system with minimal dependencies.
 		</Summary>
 	</PropertyGroup>
+	
+	<ItemGroup Condition="'$(TargetFramework)'=='net9.0'">
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[9.*, 10.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[9.*, 10.0.0)" />
+		<PackageReference Include="System.Text.Json" Version="[9.*, 10.0.0)" />
+	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[8.*-*, 9.0.0)" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[8.*-*, 9.0.0)" />
-		<PackageReference Include="System.Text.Json" Version="[8.*-*, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="System.Text.Json" Version="[8.*, 10.0.0)" />
 	</ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)'=='net7.0'">
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[7.*, 8.0.0)" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[7.*, 8.0.0)" />
-		<PackageReference Include="System.Text.Json" Version="[7.*, 8.0.0)" />
-	</ItemGroup>
-	
 	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[6.*, 8.0.0)" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[6.*, 8.0.0)" />
-		<PackageReference Include="System.Text.Json" Version="[6.*, 8.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="System.Text.Json" Version="[8.*, 10.0.0)" />
 	</ItemGroup>
 	
 	<ItemGroup Condition="'$(TargetFramework)'=='netstandard2.0'">
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[3.*, 8.0.0)" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[3.*, 8.0.0)" />
-		<PackageReference Include="System.Text.Json" Version="[5.*, 8.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="[8.*, 10.0.0)" />
+		<PackageReference Include="System.Text.Json" Version="[8.*, 10.0.0)" />
 	</ItemGroup>
 
 </Project>
