@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<AssemblyName>CloudNimble.SimpleMessageBus.Dispatch</AssemblyName>
 		<RootNamespace>CloudNimble.SimpleMessageBus.Dispatch</RootNamespace>
-		<TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0</TargetFrameworks>
+		<TargetFrameworks>netstandard2.0;net9.0;net8.0;net6.0</TargetFrameworks>
 		<DocumentationFile>$(DocumentationFile)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
 
@@ -14,30 +14,28 @@
 			A framework for reliable, distributed, scalable, cross-platform event processing on .NET. This package is 
 			installed in your Runtime app to automatically process your Queues and call your registered IMessageHandlers.
 		</Summary>
+		<PackageReleaseNotes>
+			Version 5.0:
+			- Changed the architecture to split out event stores in order to expand our capabilities. If you were using
+			  Azure or File System targets, you will need to change your Dispatch PackageReferences to use
+			  the ones with the .Azure or .FileSystem suffix instead.
+		</PackageReleaseNotes>
 	</PropertyGroup>
 
-	<ItemGroup>
-		<PackageReference Include="Azure.Security.KeyVault.Secrets" Version="[4.5.0, 5.0)" />
-		<PackageReference Include="Microsoft.Azure.WebJobs" Version="[3.0.39, 3.1.0)" />
-		<PackageReference Include="Microsoft.Azure.WebJobs.Host.Storage" Version="[5.0.0, 6.0.0)" />
-		<PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="[5.0.0, 6.0.0)" />
-		<PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Storage.Queues" Version="[5.2.*, 6.0.0)" />
+	<ItemGroup Condition="'$(TargetFramework)'=='net9.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[9.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[8.*-*, 9.0.0)" />
-	</ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)'=='net7.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[7.0.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[8.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[6.0.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[6.*, 10.0.0)" />
 	</ItemGroup>
 	
 	<ItemGroup Condition="'$(TargetFramework)'=='netstandard2.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[3.1.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[3.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup>
