@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<AssemblyName>CloudNimble.SimpleMessageBus.Hosting</AssemblyName>
 		<RootNamespace>CloudNimble.SimpleMessageBus.Hosting</RootNamespace>
-		<TargetFrameworks>net8.0;net7.0;net6.0</TargetFrameworks>
+		<TargetFrameworks>net9.0;net8.0;net6.0</TargetFrameworks>
 		<DocumentationFile>$(DocumentationFile)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
 
@@ -16,16 +16,16 @@
 		</Summary>
 	</PropertyGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[8.*-*, 9.0.0)" />
-	</ItemGroup>
-	
-	<ItemGroup Condition="'$(TargetFramework)'=='net7.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[7.0.0, 9.0.0)" />
+	<ItemGroup Condition="'$(TargetFramework)'=='net9.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[9.*, 10.0.0)" />
 	</ItemGroup>
 
+	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[8.*, 10.0.0)" />
+	</ItemGroup>
+	
 	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[6.0.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting.WindowsServices" Version="[6.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup>
