@@ -3,7 +3,7 @@
 	<PropertyGroup>
 		<AssemblyName>CloudNimble.SimpleMessageBus.Publish</AssemblyName>
 		<RootNamespace>CloudNimble.SimpleMessageBus.Publish</RootNamespace>
-		<TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0</TargetFrameworks>
+		<TargetFrameworks>netstandard2.0;net9.0;net8.0;net6.0</TargetFrameworks>
 		<DocumentationFile>$(DocumentationFile)\$(AssemblyName).xml</DocumentationFile>
 	</PropertyGroup>
 
@@ -14,31 +14,32 @@
 			A framework for reliable, distributed, scalable, cross-platform event processing on .NET. This package allows
 			client APIs and apps to publish domain messages onto the SimpleMessageBus queue while using minimal dependencies.
 		</Summary>
+		<PackageReleaseNotes>
+			Version 5.0:
+			- Changed the architecture to split out event stores in order to expand our capabilities. If you were using
+			Azure targets, you will need to change your Dispatch PackageReferences to use SimpleMessageBus.Publish.Azure
+			instead (the FileSystem publisher is still in this package).
+		</PackageReleaseNotes>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<None Include="..\.editorconfig" Link=".editorconfig" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<PackageReference Include="Azure.Security.KeyVault.Secrets" Version="[4.5.0, 5.0)" />
-		<PackageReference Include="Azure.Storage.Queues" Version="[12.*, 13.0.0)" />
+	<ItemGroup Condition="'$(TargetFramework)'=='net9.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[9.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[8.*-*, 9.0.0)" />
-	</ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)'=='net7.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[7.0.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[8.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[6.0.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[6.*, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='netstandard2.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[3.1.0, 9.0.0)" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="[3.1.0, 10.0.0)" />
 	</ItemGroup>
 
 	<ItemGroup>
