@@ -1,55 +1,57 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>net8.0;net7.0;net6.0;net472</TargetFrameworks>
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFrameworks>net9.0;net8.0;net6.0;net472</TargetFrameworks>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <None Remove="appsettings.json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Azure.WebJobs" Version="3.0.39" />
-    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="5.*" />
-    <PackageReference Include="System.Net.NameResolution" Version="4.3.0" />
-  </ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)'=='netstandard2.0' OR '$(TargetFramework)'=='net472'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="3.1.0" />
+	<ItemGroup>
+		<None Remove="appsettings.json" />
 	</ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="6.*" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="6.*" />
+	<ItemGroup>
+		<Content Include="appsettings.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</Content>
 	</ItemGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)'=='net7.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="7.*" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="7.*" />
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Azure.WebJobs" Version="3.*" />
+		<PackageReference Include="Microsoft.Azure.WebJobs.Extensions" Version="5.*" />
+		<PackageReference Include="System.Net.NameResolution" Version="4.*" />
+		<PackageReference Include="System.Private.Uri" Version="4.*" />
+	</ItemGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)'=='net9.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="9.*-*" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.*-*" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(TargetFramework)'=='net8.0'">
-		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.*-*" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.*-*" />
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.*" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.*" />
 	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\SimpleMessageBus.Core\SimpleMessageBus.Core.csproj" />
-    <ProjectReference Include="..\SimpleMessageBus.Dispatch\SimpleMessageBus.Dispatch.csproj" />
-    <ProjectReference Include="..\SimpleMessageBus.Samples.Core\SimpleMessageBus.Samples.Core.csproj" />
-    <ProjectReference Include="..\SimpleMessageBus.Samples.ExternalTriggers\SimpleMessageBus.Samples.ExternalTriggers.csproj" />
-  </ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)'=='net6.0'">
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.*" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.*" />
+	</ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' != 'net472'">
-    <ProjectReference Include="..\SimpleMessageBus.Hosting\SimpleMessageBus.Hosting.csproj" />
-  </ItemGroup>
+	<ItemGroup Condition="'$(TargetFramework)'=='netstandard2.0' OR '$(TargetFramework)'=='net472'">
+		<PackageReference Include="Microsoft.Extensions.Hosting" Version="8.*" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.*" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\SimpleMessageBus.Core\SimpleMessageBus.Core.csproj" />
+		<ProjectReference Include="..\SimpleMessageBus.Dispatch.FileSystem\SimpleMessageBus.Dispatch.FileSystem.csproj" />
+		<ProjectReference Include="..\SimpleMessageBus.Dispatch\SimpleMessageBus.Dispatch.csproj" />
+		<ProjectReference Include="..\SimpleMessageBus.Samples.Core\SimpleMessageBus.Samples.Core.csproj" />
+		<ProjectReference Include="..\SimpleMessageBus.Samples.ExternalTriggers\SimpleMessageBus.Samples.ExternalTriggers.csproj" />
+	</ItemGroup>
+
+	<ItemGroup Condition="'$(TargetFramework)' != 'net472'">
+		<ProjectReference Include="..\SimpleMessageBus.Hosting\SimpleMessageBus.Hosting.csproj" />
+	</ItemGroup>
 
 </Project>
